Componentização do NumberCase
