+++ src/components/NumberCases.vue
@@ -1,78 +1,6 @@
 <template>
   <v-container>
-    <div class="container">
-      <div class="error-alert" v-if="errors.length">
-        <v-alert
-          type="error"
-          elevation="8"
-          outlined
-          dense
-          v-for="error in errors"
-          :key="error"
-        >
-          {{ error }}
-        </v-alert>
-      </div>
-      <div class="flex">
-        <div class="flex items-center">
-          <strong class="date"> Período: </strong>
-          <div class="datepicker">
-            <datepicker
-              :calendar-button="true"
-              :clear-button="true"
-              :full-month-name="true"
-              placeholder="Data Inicial*"
-              v-model="initialDate"
-              :format="customFormatterDate"
-              :language="ptBR"
-              min="0"
-            >
-            </datepicker>
-          </div>
-        </div>
-
-        <div class="flex items-center">
-          <strong>Até</strong>
-          <div class="datepicker">
-            <datepicker
-              :calendar-button="true"
-              :clear-button="true"
-              :full-month-name="true"
-              placeholder="Data Final*"
-              v-model="finalDate"
-              :format="customFormatterDate"
-              :language="ptBR"
-              min="0"
-            >
-            </datepicker>
-          </div>
-        </div>
-      </div>
-      <br />
-
-      <v-combobox
-        class="type-of-occurrence"
-        v-model="selectedType"
-        :items="types"
-        label="Tipo de ocorrência*"
-      ></v-combobox>
-
-      <v-combobox
-        multiple
-        v-if="selectedType == 'Denúncias'"
-        class="type-complaint"
-        v-model="selectedTypeComplaint"
-        :items="typesComplaints"
-        label="Tipo de denúncia*"
-      ></v-combobox>
-
-      <div class="buttom">
-        <v-btn @click="search">Buscar</v-btn>
-        <v-btn @click="cleaner">Limpar</v-btn>
-      </div>
-      <br />
-      <small>* Campos obrigatórios</small>
-    </div>
+    <Form @my-alerts="getAlerts" @my-complaints="getComplaints" @my-clean="cleanLoading"/>
 
     <div class="container-chart">
       <div class="chart-alerts" v-if="isLoadedAlert">
@@ -125,177 +53,79 @@
 </template>
 
 <script>
-import Datepicker from "vuejs-datepicker";
-import { ptBR } from "vuejs-datepicker/dist/locale";
-import moment from "moment";
 import BarChart from "./BarChart.vue";
 import DoughnutChart from "./DoughnutChart.vue";
+import Form from './Form.vue';
 
 export default {
   name: "numberCases",
+  
   components: {
-    Datepicker,
     BarChart,
     DoughnutChart,
-  },
-  data() {
-    return {
-      ptBR: ptBR,
-      initialDate: "",
-      finalDate: "",
-      token: "",
-      selectedType: "",
-      types: ["Alertas", "Denúncias"],
-      selectedTypeComplaint: [],
-      typesComplaints: [
-        "Todas",
-        "Física",
-        "Moral",
-        "Sexual",
-        "Patrimonial",
-        "Psicológica",
-        "Verbal",
-      ],
-      chartsData: {},
-      alertsByDates: {},
-      alertsByDistricts: {},
-      complaintsByDates: {},
-      complaintsByDistricts: {},
-      complaintsByTypes: {},
-      errors: [],
-      isLoadedAlert: false,
-      isLoadedComplaint: false,
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-      },
-    };
+    Form
   },
 
   mounted() {
     this.token = sessionStorage.getItem('token')
   },
 
+  data() {
+    return {
+      token: "",
+      isLoadedAlert: false,
+      isLoadedComplaint: false,
+      alertsByDates: {},
+      alertsByDistricts: {},
+      complaintsByDates: {},
+      complaintsByDistricts: {},
+      complaintsByTypes: {},
+    };
+  },
+
   methods: {
-    customFormatterDate(date) {
-      return moment(date).format("DD/MM/YYYY");
-    },
-
-    customFormatterDateDayMonth(date) {
-      return moment(date).format("DD/MM");
-    },
-
-    logout() {
-      sessionStorage.removeItem("token");
-      this.$router.replace("/login");
-    },
 
     async getAlerts(date) {
       this.$api.defaults.headers.common[
           "Authorization"
       ] = `Bearer ${this.token}`;
       await this.$api
-        .get(`/alert?init=${date.init}&final=${date.final}`)
-        .then((response) => {
+      .get(`/alert?init=${date.init}&final=${date.final}`)
+      .then((response) => {
           this.alertsByDates = response.data.Date;
           this.alertsByDistricts = response.data.District;
           this.isLoadedAlert = true;
           this.isLoadedComplaint = false;
-        })
+      })
     },
 
-    async getComplaints(date) {
-      const type = this.selectedTypeComplaint == "Todas" ? "all" : this.selectedTypeComplaint;
-      
-      this.$api.defaults.headers.common[
+    async getComplaints(date, typeComplaint) {
+        const type = typeComplaint == "Todas" ? "all" : typeComplaint;
+        
+        this.$api.defaults.headers.common[
         "Authorization"
-      ] = `Bearer ${this.token}`;
-      await this.$api
+        ] = `Bearer ${this.token}`;
+        await this.$api
         .get(`/complaint?init=${date.init}&final=${date.final}&type=${type}`)
         .then((response) => {
-          this.complaintsByDates = response.data.Date;
-          this.complaintsByDistricts = response.data.District;
-          this.complaintsByTypes = response.data.Type;
-          this.isLoadedComplaint = true;
-          this.isLoadedAlert = false;
+            this.complaintsByDates = response.data.Date;
+            this.complaintsByDistricts = response.data.District;
+            this.complaintsByTypes = response.data.Type;
+            this.isLoadedComplaint = true;
+            this.isLoadedAlert = false;
         })
     },
 
-    search() {
-      this.errors = [];
-
-      if (!this.initialDate || !this.finalDate || !this.selectedType) {
-        this.errors.push("Por favor, preencha os campos corretamentes!");
-        return;
-      } else if (this.selectedType === "Denúncias" && this.selectedTypeComplaint.length === 0){
-        this.errors.push("Por favor, escolha o Tipo de denúncia!");
-        return;
-      } else if (this.finalDate < this.initialDate) {
-        this.errors.push("Por favor, informe a data final maior que data inicial!");
-        return;
-      } else {
-        const dates = {
-          init: moment(this.initialDate).format("YYYY-MM-DD"),
-          final: moment(this.finalDate).format("YYYY-MM-DD"),
-        };
-
-        this.selectedType === "Alertas"
-          ? this.getAlerts(dates)
-          : this.getComplaints(dates);
-        }
-    },
-    
-    cleaner() {
-      this.errors = [];
-      this.initialDate = "";
-      this.finalDate = "";
-      this.selectedType = "";
+    cleanLoading(){
       this.isLoadedAlert = false;
       this.isLoadedComplaint = false;
-    },
-  },
+    }
+  }
+
 };
 </script>
 
 <style scoped>
-.error-alert {
-  display: flex;
-  justify-content: space-evenly;
-}
-.container {
-  display: grid;
-  justify-content: center;
-}
-.flex {
-  display: flex;
-}
-
-.items-center {
-  align-items: center;
-}
-
-.datepicker {
-  margin-left: 10px;
-  margin-right: 10px;
-  border-bottom: solid 1px #9e9e9e;
-  padding: 5px;
-}
-
-.typeSelect {
-  margin-left: 10px;
-  border: solid 1px;
-}
-
-#types {
-  display: flex;
-  align-items: center;
-}
-
-.buttom {
-  margin: 1rem;
-  display: flex;
-  justify-content: space-evenly;
-}
 
 .showMap {
   display: grid;
@@ -327,40 +157,5 @@ export default {
   border: solid 1px #555;
   box-shadow: 0 0 10px rgba(0, 0, 0, 0.6);
   -webkit-box-shadow: 0 0 10px rgba(0, 0, 0, 0.6);
-}
-
-@media only screen and (max-width: 765px) {
-  .items-center {
-    padding: 1rem 0 1rem 0;
-  }
-
-  .items-center strong {
-    display: none;
-  }
-
-  .datepicker {
-    font-size: 15px;
-    width: 75vw;
-  }
-
-  .type-of-occurrence {
-    margin: -1rem 0 0 0.75rem;
-    align-content: center;
-    width: 75vw;
-  }
-
-  .type-complaint {
-    margin: 0 0 0 0.75rem;
-    align-content: center;
-    width: 75vw;
-  }
-
-  .flex {
-    display: grid;
-  }
-
-  strong {
-    font-size: 15px;
-  }
 }
 </style>
