+++ src/views/Home.vue
@@ -51,6 +51,7 @@ export default {
     },
   },
 };
+
 </script>
 
 <style >
